Split tests so that they are compatible with scala native
