+++ core/src/test/scalajvm/sttp/tapir/internal/RichEndpointOutputTest.scala
@@ -5,6 +5,7 @@ import org.scalatest.matchers.should.Matchers
 import sttp.model.MediaType
 import sttp.tapir._
 
+// TODO: move to shared test sources when https://github.com/softwaremill/sttp-model/issues/188 is fixed
 class RichEndpointOutputTest extends AnyFlatSpec with Matchers {
   "output media type" should "match content type with lower and upper case charset" in {
     val o = endpoint.put
